Add Sum of Square Numbers

+++ Sum_of_Square_Numbers.cpp
@@ -5,10 +5,9 @@ public:
         return false;
     }
     bool judgeSquareSum(int c) {
-        for(long long i=0; i*i <= c;i++){
+        for(int i=0; i*i <= c;i++){
             if(check(c-(i*i))) return true;
         }
         return false;
     }
 };
-// Title: Sum of Square Numbers
